Buffer safety and sign extension fixes (thanks Coverity Scan).

+++ libparse/binio.c
@@ -69,7 +69,7 @@ get_lsb_long(
   retval  = *((*bufpp)++);
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++) << 16;
-  retval |= *((*bufpp)++) << 24;
+  retval |= (u_long)*((*bufpp)++) << 24;
 
   return retval;
 }
@@ -116,7 +116,7 @@ get_msb_long(
 {
   long retval;
 
-  retval  = *((*bufpp)++) << 24;
+  retval  = (u_long)*((*bufpp)++) << 24;
   retval |= *((*bufpp)++) << 16;
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++);
